Added support for sub-wheels to `pyqt-qt-wheel`
The (undocumented) `pyqt-qt-wheel` utility now supports the splitting of a
project's wheel into the main wheel and a sub-wheel.  Specifying
`--sub-wheel generate` will generate the sub-wheel, and
`--sub-wheel exclude` will generate the main wheel (ie. a normal wheel
without the contents of the sub-wheel).  By default a normal wheel is
generated.